Making TS work when you don't have a delayed_jobs table

+++ lib/thinking_sphinx/deltas/delayed_delta/job.rb
@@ -6,7 +6,9 @@ module ThinkingSphinx
       end
       
       def self.cancel_thinking_sphinx_jobs
-        delete_all("handler LIKE '--- !ruby/object:ThinkingSphinx::Deltas::%'")
+        if connection.tables.include?("delayed_jobs")
+          delete_all("handler LIKE '--- !ruby/object:ThinkingSphinx::Deltas::%'")
+        end
       end
 
       private
